move test_hierarchy_array_object to test_array_simple
