+++ tests/test_cases/test_array_simple/test_array_simple.py
@@ -10,6 +10,7 @@ import pytest
 
 import cocotb
 from cocotb.clock import Clock
+from cocotb.handle import HierarchyObject
 from cocotb.triggers import Timer
 
 tlog = logging.getLogger("cocotb.test")
@@ -160,6 +161,25 @@ async def test_struct_unpacked(dut):
     dut.inout_if.a_in.value = 0
     await Timer(1000, "ns")
     _check_value(tlog, dut.inout_if.a_in, 0)
+
+
+@cocotb.test()
+async def test_hierarchy_array_generic_typing(dut):
+    """Test that HierarchyArrayObject generic typing works correctly"""
+    tlog = logging.getLogger("cocotb.test")
+    arr_gen = dut.arr
+    arr_element_1, arr_element_2 = arr_gen[1], arr_gen[2]
+    assert isinstance(arr_element_1, HierarchyObject)
+    assert isinstance(arr_element_2, HierarchyObject)
+
+    for element in arr_gen:
+        assert isinstance(element, HierarchyObject)
+        tlog.info("Iteration element type: %s", type(element).__name__)
+        break
+
+    first_element = next(iter(arr_gen))
+    assert isinstance(first_element, HierarchyObject)
+    tlog.info("HierarchyArrayObject[HierarchyObject] works correctly")
 
 
 @cocotb.test()
